Update pyface in the conda environment for mayavi
The anaconda virtual environment installs mayavi with a set of
dependencies. There issue is that the fixed version of pyface that
matches the anaconda installation conflicts with sphinx and pygments for
syntax highlighting. Thus pyface has to updated again and individually
to bypass the fixed version dependencies imposed by anaconda in mayavi.