Automate touching of non-existing backrefences
This is to purge the output of sphinx as it complains when using autodoc
autosummary that includes backreferences to each function and class even
if it does not exist.

Touching the backreference files automatically to avoid inclusion error
is only relevant when using autodoc. Because auto doc tries to include
backreferences to every function/class.

If the user does not configure for autodoc, then this step is not needed
and not even possible as it is linked to the
autodoc-processing-docstring step.