Pyface pip update to pass travis Mayavi build
The conda virtualenv for Mayavi forces a version of pyface that clashes
with sphinx. The manual update of pyface within conda is no longer
enough to update to a new version that does not clash with sphinx. Thus
the update is forced through pip.

Mayavi is an experimentally supported use case of Sphinx-Gallery