Rename the control variable to run examples
Because of the need to stop running examples after they fail and to
simplify the control flow logic on when to parse the file only. All were
joined in a execution global private variable __example_has_failed__.

Now this is renamed to __sphx_glr_run_example__ to better fit the
reading of the code.