[CSS] Code link tooltip is left aligned with code
Code link tooltips used to be shifted to the left to give the
impression they are centered with the link one hovers. Nevertheless
this floating tooltip collides with the sidebar in the Scikit-learn
documentation.

As a general goal our CSS shall not collide with other projects
templates. Thus this tooltip shall be contained with the codeblock
where the links are.