[MRG] Documentation Auto API (#151)
* [DOC] Organization

- Fixes to Sphinx Headers

Apparently Sphinx does not like the overline for the headers. Especially
when using the auto docs. Also it does not look good in the Jupyter
notebooks. Removing the overline does not have any negative effect on
the html heading.

- Fixes to level documentation

* Use backreferences_dir variable

Deprecationwarning on mod_examples_dir new of default

* [DOC] including example backreferences

Default to None example back-references

Move examples backreferences sources

* [DOC] Explain autosummary

* Information on new backreferences config

* Use of intersphinx to document numpy, scipy & matplotlib

* Backreferences fail message don't raise exception