BUG: Fix reading with non-utf-8 default encoding
Standardize the utf-8 handling a bit more by:
- open in binary mode
- use a decode('utf-8') whenever we read
- don't try to check for bytes anywhere else

This fixes a problem where open() defaults to text mode, invoking the
default system encoding. For files written in utf-8, this results in
decode errors on systems where the default encoding is something else. This
is especially problematic on windows where the default is cp1252.