Don't create spurious examples/ directory. (#367)
When using the "separate source and build" layout, sphinx-gallery
creates a spurious (and ultimately empty) "examples/" directory next to
the Makefile, whereas the correct place for that directory is
"source/examples".  It is in fact not necessary to pre-create that
directory, because generate_dir_rst will create the correct "examples"
folder.