Don't show "Total running time" if SOURCE_DATE_EPOCH is set (Closes: #390)
Whilst working on the Reproducible Builds effort [0], we noticed that
sphinx-gallery could not be built reproducibly.

This patch removes the "Total running time of the script" messages from the
documentation build if the SOURCE_DATE_EPOCH [1] environment variable is
exported in the environment.

This was originally filed in Debian as #901307 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/901307