Make md5 hash independent of platform line endings
Currently md5 hash for Python files depends on what platform file is on,
as file endings will differ. This change reads the Python files in the
hashing function with universal line endings, such md5 is independent of
platform being used.

Surrogate escapes are used, to ensure that it can handle any errors
without data loss.

This is useful when adding a complex or long running examples into a
repository that may be used across platforms.