ENH Add support for image links and data URIs for notebooks
Currently, when notebooks are generated, the links to images found in
rST remain unmodified, meaning images expected to be either local or
found on web server hosting documentation are broken.

This change allows inclusion of images via data URIs, or prefixing the
image links in the notebooks, for example to server hosting
documentation.

This also changes the markdown output to use HTML `img` tag, and such
also supports the width and height options.