DOC: clarify advanced doc wrt referencing examples
As discussed via https://github.com/sphinx-gallery/sphinx-gallery/issues/805, the documentation could be structured a bit better to make it easier to find how to link examples. This PR suggests adding a new sub-heading *Referencing Examples* so that the term *reference* explicitly shows up in the table of contents. 

Two things to discuss:
* I'm not sure yet whether the heading should be *referencing* or *linking*. 
* Is it possible to link to heading in examples, too? Is there a special syntax for that? If yes, I could extend this PR to include it.