Fix broken matplotlib link
A URL is better than a sphinx ref. On the matplotlib side, we at least try to add
redirects when we change URLs. And even if that's not the case, you'll get
to the matplotlib main page. We have now way of knowing whether external sites
reference our doc labels, so they can get broken without notice, and when the do
it's completely dead.