Fix "Out" layout for pydata-sphinx-theme (#886)
* Fix "Out" layout for pydata-sphinx-theme

Follow up to #484, which got misaligned again, likely through pydata-sphinx-theme.

Note: This is a purely empirical suggestion based on the matplotlib docs, e.g.:
https://matplotlib.org/devdocs/tutorials/colors/colormap-manipulation.html

I don't know how to find out whether this is generally ok.

* Fix broken matplotlib link

A URL is better than a sphinx ref. On the matplotlib side, we at least try to add
redirects when we change URLs. And even if that's not the case, you'll get
to the matplotlib main page. We have now way of knowing whether external sites
reference our doc labels, so they can get broken without notice, and when the do
it's completely dead.