Use pseudo-elements for 'Out:' prefixing (#896)
There's still an arbitrary padding, but now it goes on the added content
instead of the existing `pre`. Without the negative margin, this
prevents the `pre` from accidentally overlapping previous content. Some
additional tweaks are necessary to work with the pydata sphinx theme,
but it appears to work for the rtd theme as well.