Fix reset_argv example
The example of how to use `reset_argv` is wrong, or at least misleading on two counts.
One, it must return an iterable in all cases, so there should be a default case
returning `[]`. Two, the value of `script_vars['src_file']` is an absolute path, not
just a filename, so just `sample1.py` won't ever be matched.